ps-repo-manager/pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "ps-repo-manager"
version = "0.8.2"
description = "Tools for managing Polestar Repos."
authors = [ "edkirk30 <edward.kirk@polestarglobal.com>",]
[[tool.poetry.source]]
name = "ps-python"
url = "https://polestar-tools-324252367609.d.codeartifact.us-east-1.amazonaws.com/pypi/ps-python/simple/"
secondary = true

[[tool.poetry.packages]]
include = "repomanager"

[tool.isort]
profile = "black"
src_paths = [ ".",]
line_length = 100

[tool.black]
line-length = 100
target-version = [ "py36", "py37", "py38",]

[tool.poetry.dependencies]
python = "^3.6"
typer = "^0.7.0"
cerberus = "^1.3.4"
pyyaml = ">=3.1"
gitpython = "^3.1.18"
semver = "^2.13.0"
toml = "^0.10.2"
path = "^16.2.0"
[[tool.poetry.dependencies.importlib-metadata]]
version = "4.13.0"
python = ">=3.7"

[[tool.poetry.dependencies.importlib-metadata]]
version = "4.8.3"
python = "<3.7"

[[tool.poetry.dependencies.click]]
version = "8.1.3"
python = ">=3.7"

[[tool.poetry.dependencies.click]]
version = "8.0.4"
python = "<3.7"

[tool.poetry.scripts]
repo-manager = "repomanager.cli:app"
ps-repo-manager = "repomanager.cli:app"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
nox = "^2022.1.7"
pexpect = "^4.8.0"

[tool.poetry.dependencies.textual]
version = "<0.11.0"
python = ">=3.7"

[tool.poetry.dev-dependencies.textual]
version = "*"
python = ">=3.7"
extras = [ "dev",]
